修复了打包工具因引擎架构修改而产生的bug

+++ builder.py
@@ -1,11 +1,12 @@
 import os
 import shutil
 
-#新建一个key,以告诉编译器需要保留pyd文件
-with open("../linpg/building_key.txt", "r", encoding='utf-8') as f: pass
-
-#编译所有文件
-os.system("python compiler.py build_ext --inplace")
+#先编译py文件
+if not os.path.exists("src") or input("Do you want to compile everything:").lower() == "y":
+    #新建一个key,以告诉编译器需要保留pyd文件
+    with open("../linpg/building_key.txt", "w", encoding='utf-8') as f: pass
+    #编译所有文件
+    os.system("python compiler.py build_ext --inplace")
 
 #升级build工具
 os.system("python -m pip install --upgrade build")
@@ -20,6 +21,6 @@ os.system("python -m pip install --user --upgrade twine")
 os.system("twine upload dist/*")
 
 #删除不需要的文件
-folders_need_remove = ["dist","linpg.egg-info","Save","build"]
+folders_need_remove = ["dist","Save","build"]
 for path in folders_need_remove:
-    shutil.rmtree(path)
+    if os.path.exists(path): shutil.rmtree(path)
